Sort root services into platform/tenant homes
File the services that clearly own one side of the tenancy boundary
(per services/README.md: file by the data a service reads/writes):

- tenant/: stats_service, import_service, attachments, collaboration,
  recurrence, mention_parser (+ their co-located tests)
- platform/: push_notifications, csv_export (+ test)

Root keepers verified as genuine bridges/infra: oidc_sync (syncs tenant
initiative membership from IdP claims), storage (blob infra used by db
layer + main), webhook_target_url (SSRF guard shared by tenant webhooks
and AI settings), plus everything README already lists.

Also split the one two-axis endpoint file: the platform AI-settings
router (owner-only, mounted top-level) moves from
tenant_endpoints/ai_settings.py to platform_endpoints/ai_settings.py.
Route paths, tags, and handler names are unchanged — the exported
OpenAPI spec is byte-identical, so no generated-client churn. Delete
the empty vestigial guild_endpoints/ and public_endpoints/ dirs.
