+++ backend/app/services/tenant/recurrence_test.py
@@ -10,7 +10,7 @@ Tests the recurrence calculation logic including:
 from datetime import datetime, timezone
 
 from app.schemas.tenant.task import TaskRecurrence
-from app.services.recurrence import get_next_due_date
+from app.services.tenant.recurrence import get_next_due_date
 
 
 class TestDailyRecurrence:
