+++ backend/app/services/tenant/stats_service.py
